+++ anyblok/model/common.py
@@ -5,26 +5,22 @@
 # This Source Code Form is subject to the terms of the Mozilla Public License,
 # v. 2.0. If a copy of the MPL was not distributed with this file,You can
 # obtain one at http://mozilla.org/MPL/2.0/.
-from anyblok.field import Field, FieldException
+from .factory import has_sql_fields  # noqa
+from .factory import ModelFactory, ViewFactory, ModelFactoryException
+import warnings
 
-MODEL = "MODEL"
-VIEW = "VIEW"
+MODEL = 'MODEL'
+VIEW = 'VIEW'
 
 
-def has_sql_fields(bases):
-    """ Tells whether the model as field or not
-
-    :param bases: list of Model's Class
-    :rtype: boolean
-    """
-    for base in bases:
-        for p in base.__dict__.keys():
-            try:
-                if hasattr(getattr(base, p), '__class__'):
-                    if Field in getattr(base, p).__class__.__mro__:
-                        return True
-            except FieldException:
-                # field function case already computed
-                return True
-
-    return False
+def get_factory(type_):
+    warnings.warn(
+        "type attribute is deprecated to define a type"
+        " of model, you have to use the factory",
+        DeprecationWarning, stacklevel=2)
+    if type_ == MODEL:
+        return ModelFactory
+    elif type_ == VIEW:
+        return ViewFactory
+    else:
+        raise ModelFactoryException("The type %r is unknown" % type_)